+++ CHANGELOG.md
@@ -1,14 +1,19 @@
 # CHANGELOG
 
-# 2022-01-15
+
+## 2022-04-10
+* Defined a color set for each season (winter, spring, summer, fall)
+* Added code to change the color set based on the season as determined by the current date.
+
+## 2022-01-15
 * Changed to "Winter" (cold) colors
 * Show the time always; no effects
 
-# 2021-12-20
+## 2021-12-20
 * Added Effects
 * Re-programmed to display the time every 15 minutes, otherwise it displays a
   rainbow color wheel.
 
-# 2020-12-16
+## 2020-12-16
 * Updated to work with latest PlatformIO
 * Updated light color for Christmas